# src/main/resources/static/swagger/yaml/data.yaml
swagger: "2.0"
info:
  description: "日升木业后台接口"
  version: "1.0.0"
  title: "sunup"
...
schemes:
- "http"
produces:
- "application/json"
securityDefinitions:
  token:
    type: apiKey
    name: Cookie
    in: header
security:
- token: []
paths:
...
  /project/user:
    get:
      tags:
        - "ProjectUser"
      summary: "获取用户信息"
      operationId: getUser
      parameters:
        - in: query
          name: id
          type: string
      responses:
        200:
          schema:
            $ref: '#/definitions/UserDTO'
          description: "用户列表"

    delete:
      summary: 删除用户
      operationId: deleteById
      tags:
        - "ProjectUser"
      parameters:
        - in: query
          type: string
          name: id
          description: 删除的ID
      responses:
        200:
          description: "删除成功"
    put:
      tags:
        - "ProjectUser"
      summary: 修改用户信息
      operationId: updateUser
      parameters:
        - name: "user"
          in:  body
          required: true
          schema:
            $ref: '#/definitions/UserDTO'

      responses:
        200:
          description: "修改成功"
...
  /project/product/designer:
    post:
      summary: "添加设计师"
      operationId: addDesigner
      tags:
        - "ProjectDesigner"
      parameters:
        - in: body
          name: designerDto
          schema:
            $ref: '#/definitions/DesignerDTO'
          required: true
      responses:
        200:
          description: "新增成功"
    delete:
      tags:
        - "ProjectDesigner"
      summary: "删除设计师"
      operationId: deleteById
      parameters:
        - in: query
          name: id
          type: string
          required: true
      responses:
        200:
          description: "删除成功"
    put:
      tags:
        - "ProjectDesigner"
      summary: "修改设计师信息"
      operationId: modifyDesignerInfo
      parameters:
        - in: body
          name: designerDto
          schema:
            $ref: '#/definitions/DesignerDTO'
          required: true
      responses:
        200:
          description: "修改成功"
    get:
      tags:
        - "ProjectDesigner"
      summary: "查看设计师详细信息"
      operationId: getDesignerInfo
      parameters:
        - in: query
          name: id
          type: string
          required: true
      responses:
        200:
          schema:
            $ref: '#/definitions/DesignerDTO'
          description: "设计师内容信息"
  /project/product/designer/queryList:
    get:
      tags:
        - "ProjectDesigner"
      summary: "查看设计师列表"
      operationId: getDesignerList
      parameters:
        - in: query
          name: key
          type: string
        - in: query
          name: pageIndex
          type: integer
          description: 当前页数
          default: 0
        - in: query
          name: pageSize
          type: integer
          description: 页面大小
          default: 3
      responses:
        200:
          schema:
            type: array
            items:
              $ref: '#/definitions/DesignerDTO'
          description: "设计师列表"
  /project/product/designer/getReels:
    get:
      summary: 获取设计师作品列表
      operationId: getReelsById
      tags:
        - "ProjectDesigner"
      parameters:
        - in: query
          type: string
          name: id
          description: 设计师ID
          required: true
      responses:
        200:
          schema:
            type: array
            items:
              type: string
          description: 作品列表
  /project/other/banner:
    get:
      summary: 获取banner图信息
      operationId: getBannerById
      tags:
        - "ProjectBanner"
      parameters:
        - in: query
          name: id
          required: true
          type: string
      responses:
        200:
          schema:
            $ref: '#/definitions/BannerDTO'
          description: banner图内容
    delete:
      summary: 删除banner图信息
      operationId: deleteBannerById
      tags:
        - "ProjectBanner"
      parameters:
        - in: query
          name: id
          required: true
          type: string
      responses:
        200:
          description: 删除成功
    post:
      summary: 新增banner图
      operationId: addBanner
      tags:
        - "ProjectBanner"
      parameters:
        - in: body
          name: banner
          required: true
          schema:
            $ref: '#/definitions/BannerDTO'
          description: 序号，图片必须有
      responses:
        200:
          description: 新增成功
    put:
      summary: 修改banner图
      operationId: updateBanner
      tags:
        - "ProjectBanner"
      parameters:
        - in: body
          name: banner
          required: true
          schema:
            $ref: '#/definitions/BannerDTO'
      responses:
        200:
          description: 修改成功
  /project/other/banner/queryList:
    get:
      tags:
        - "ProjectBanner"
      summary: "Banner图列表"
      operationId: getBannerList
      parameters:
        - in: query
          name: pageIndex
          type: integer
          description: 当前页数
          default: 0
        - in: query
          name: pageSize
          type: integer
          description: 页面大小
          default: 3
      responses:
        200:
          schema:
            type: array
            items:
              $ref: '#/definitions/BannerDTO'
          description: "Banner图列表"
  /project/design/adorn:
    post:
      summary: 装饰设计-添加
      tags:
        - "AdornDesign"
      operationId: addAdorn
      parameters:
        - in: body
          name: adornDto
          required: true
          description: 装饰设计内容
          schema:
            $ref: '#/definitions/AdornDesignDTO'
            required:
              - name
      responses:
        200:
          description: 添加成功
    put:
      tags:
        - "AdornDesign"
      summary: 修改装饰设计内容
      operationId: updateAdorn
      parameters:
        - in: body
          name: adornDto
          required: true
          description: 装饰设计内容
          schema:
            $ref: '#/definitions/AdornDesignDTO'
            required:
              - name
      responses:
        200:
          description: 修改成功
    delete:
      tags:
        - "AdornDesign"
      summary: 删除装饰内容
      operationId: deleteById
      parameters:
        - in: query
          name: id
          type: string
          required: true

      responses:
        200:
          description: 删除成功
    get:
      tags:
        - "AdornDesign"
      summary: 装饰设计内容信息
      operationId: getAdornInfo
      parameters:
        - in: query
          name: id
          type: string
          required: true
          description: 唯一ID
      responses:
        200:
          schema:
            $ref: '#/definitions/AdornDesignDTO'
          description: "获取成功"
...
  /project/design/product:
    post:
      summary: 产品设计-添加
      tags:
        - "ProductDesign"
      operationId: addProductDesign
      parameters:
        - in: body
          name: productDto
          required: true
          description: 产品设计内容
          schema:
            $ref: '#/definitions/ProductDesignDTO'
      responses:
        200:
          description: 添加成功
    put:
      tags:
        - "ProductDesign"
      summary: 修改装饰设计内容
      operationId: updateProductDesign
      parameters:
        - in: body
          name: designDto
          required: true
          description: 产品设计内容
          schema:
            $ref: '#/definitions/ProductDesignDTO'
      responses:
        200:
          description: 修改成功
    delete:
      tags:
        - "ProductDesign"
      summary: 删除装饰内容
      operationId: deleteById
      parameters:
        - in: query
          name: id
          type: string
          required: true
      responses:
        200:
          description: 删除成功
    get:
      tags:
        - "ProductDesign"
      summary: 产品设计内容信息
      operationId: getProductDesignDTOInfo
      parameters:
        - in: query
          name: id
          type: string
          required: true
          description: 唯一ID
      responses:
        200:
          schema:
            $ref: '#/definitions/ProductDesignDTO'
          description: "获取成功"
...
definitions:
  AdornDesignDTO:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
        description:  描述
      rank:
        type: integer
        description: 排序
      lordImg:
        type: string
        description: 主图
      imgUrls:
        type: array
        items:
          type: string
          description: 图片集
  ProductDesignDTO:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
        description: 中文描述
      rank:
        type: integer
        description: 排序
      egName:
        type: string
        description: 英文描述
      lordImg:
        type: string
        description: 主图
      isLarge:
        type: integer
        description: 是否是大图
      imgUrls:
        type: array
        items:
          type: string
          description: 图片集
  BannerDTO:
    type: object
    properties:
      id:
        type: string
      title:
        type: string
        description: 标题描述
      imgUrl:
        type: string
        description: 图片地址
      rank:
        type: integer
        description: 排序
  DesignerDTO:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
        description: 设计师名
      rank:
        type: integer
        description: 排序
      sex:
        type: string
        description: 性别
      age:
        type: string
        description: 年龄
      photo:
        type: string
        description: 个人照片
      introduce:
        type: string
        description: 个人介绍
      imgReels:
        type: array
        items:
          type: string
        description: 设计师作品集图片列表
...
  UserDTO:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
        description: 用户名
      password:
        type: string
        description: 用户密码
      email:
          type: string
          description: 邮箱
      phone:
          type: string
          description: 手机号
...